lab07c/python/requirements.txt
# Python Requirements for Lab 07c: Advanced LangGraph with Multi-Database State Management
# Install with: pip install -r requirements.txt

# Core Snowflake connectivity (inherited from lab05)
snowflake-connector-python==3.6.0
pandas>=2.2.0
numpy>=1.26.0

# LangChain core framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# LangGraph framework (next-generation agent framework)
langgraph>=0.1.0

# Database connectivity for state management
psycopg2-binary>=2.9.0  # PostgreSQL
pyodbc>=4.0.39          # SQL Server/Azure SQL
azure-cosmos>=4.5.0     # Azure Cosmos DB
aiosqlite>=0.19.0       # Async SQLite

# Azure AD Authentication
azure-identity>=1.15.0  # Azure AD integrated authentication
msal>=1.24.0            # Microsoft Authentication Library
cryptography>=41.0.0    # For token encryption/decryption

# Azure OpenAI integration
openai>=1.7.0

# Environment management
python-dotenv==1.0.0

# HTTP requests for currency conversion API
requests>=2.28.0

# File handling and code interpretation
pypdf2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.20
openpyxl==3.1.2

# Additional utilities
requests==2.31.0
sqlparse==0.4.4
pydantic>=2.5.0
tiktoken>=0.5.0

# Web framework (optional for advanced examples)
flask==3.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Development and testing
jupyter==1.0.0
ipykernel==6.25.0